+++ HealthManagement/HealthManagement-ejb/src/java/ejbs/FaqBean.java
@@ -1,13 +1,20 @@
 package ejbs;
 
 import dtos.FaqDTO;
+import entities.Caregiver;
 import entities.FAQ;
+import entities.Need;
+import entities.Proceeding;
+import exceptions.EntityDoesNotExistException;
+import exceptions.MyConstraintViolationException;
+import exceptions.Utils;
 import java.util.ArrayList;
 import java.util.List;
 import javax.ejb.EJBException;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.validation.ConstraintViolationException;
 
 @Stateless
 public class FaqBean {
@@ -15,22 +22,25 @@ public class FaqBean {
     @PersistenceContext
     EntityManager em;
 
-    /* Mesmo parametros */
-    public void create(String description, String question, String answer) {
+    public void create(String description, String question, String answer)
+            throws MyConstraintViolationException {
         try {
             FAQ faq = new FAQ(description, question, answer);
 
             em.persist(faq);
+        } catch (ConstraintViolationException e) {
+            throw new MyConstraintViolationException(Utils.getConstraintViolationMessages(e));
         } catch (EJBException e) {
             throw new EJBException(e.getMessage());
         }
     }
 
-    public void update(int id, String description, String question, String answer) {
+    public void update(int id, String description, String question, String answer) 
+            throws EntityDoesNotExistException, MyConstraintViolationException {
         try {
             FAQ faq = em.find(FAQ.class, id);
             if (faq == null) {
-                return;
+                throw new EntityDoesNotExistException("There is no FAQ with that id.");
             }
 
             faq.setDescription(description);
@@ -38,36 +48,62 @@ public class FaqBean {
             faq.setAnswer(answer);
 
             em.merge(faq);
+        } catch (EntityDoesNotExistException e) {
+            throw e;
+        } catch (ConstraintViolationException e) {
+            throw new MyConstraintViolationException(Utils.getConstraintViolationMessages(e));
         } catch (EJBException e) {
             throw new EJBException(e.getMessage());
         }
     }
-
-    public FAQ getFaq(int id) {
-        try {
-            FAQ faq = em.find(FAQ.class, id);
-
-            return faq;
-        } catch (EJBException e) {
-            throw new EJBException(e.getMessage());
-        }
-    }
-
-    public void remove(int id) {
-        try {
-            FAQ faq = em.find(FAQ.class, id);
-
-            em.remove(faq);
-        } catch (EJBException e) {
-            throw new EJBException(e.getMessage());
-        }
-    }
-
+    
     public List<FaqDTO> getAll() {
         try {
             List<FAQ> faqs = (List<FAQ>) em.createNamedQuery("getAllFAQs").getResultList();
 
             return faqsToDTOs(faqs);
+        } catch (EJBException e) {
+            throw new EJBException(e.getMessage());
+        }
+    }
+
+    public FaqDTO getFaq(int id) throws EntityDoesNotExistException {
+        try {
+            FAQ faq = em.find(FAQ.class, id);
+            if (faq == null) {
+                throw new EntityDoesNotExistException("There is no FAQ with that id.");
+            }
+            
+            return faqToDTO(faq);
+        } catch (EntityDoesNotExistException e) {
+            throw e;
+        } catch (EJBException e) {
+            throw new EJBException(e.getMessage());
+        }
+    }
+
+    public void remove(int id) throws EntityDoesNotExistException {
+        try {
+            FAQ faq = em.find(FAQ.class, id);
+            if (faq == null) {
+                throw new EntityDoesNotExistException("There is no FAQ with that id.");
+            }
+
+            for (Caregiver caregiver : faq.getCaregivers()) {
+                caregiver.removeMaterial(faq);
+            }
+            
+            for (Need need : faq.getNeeds()) {
+                need.removeMaterial(faq);
+            }
+            
+            for (Proceeding proceeding : faq.getProceedings()) {
+                proceeding.setMaterial(null);
+            }
+            
+            em.remove(faq);
+        } catch (EntityDoesNotExistException e) {
+            throw e;    
         } catch (EJBException e) {
             throw new EJBException(e.getMessage());
         }
